🐛 Fix unzipping of extensions that already exist
This role works great for the initial installation of extensions, but
running a second time leads to this task failing with an
UnarchiveError with a message of the form "Cannot change group
ownership of <file path> to <user>, as user <user>"

Ideally the unarchive module would just move past files that already
exist in the destination (at least when the file hash is
identical...), but it doesn't seem to.

Since we know that every extension will have a metadata.json file, we
can use the 'creates' parameter to conditionally skip this unarchiving
if this file already exists.

See the docs for unarchive for more information on this parameter:
https://docs.ansible.com/ansible/latest/modules/unarchive_module.html